docs: fix docs and test suites
Tests in test-suite-java are missing from Groovy and Kotlin test-suites
This means when they are documented, the multi-language selector disappears when switching to non-java
Fixing this by migrating the tests from java into groovy and kotlin
Also, the description is missing from several configoptions (as the getter override has been javadoc decorated instead of the setter)
