+++ test-suite-groovy/build.gradle
@@ -10,11 +10,21 @@ dependencies {
     testImplementation(mnValidation.micronaut.validation)
     testImplementation(mnTest.micronaut.test.spock)
 
+    testImplementation(mnSecurity.micronaut.security)
     testImplementation(mnSerde.micronaut.serde.jackson)
     testImplementation(mn.micronaut.http.server.netty)
     testImplementation(mn.micronaut.http.client)
+    testImplementation(mn.reactor)
+
+    testImplementation(mnSecurity.micronaut.security)
+    testImplementation(mnSerde.micronaut.serde.api)
+    testImplementation(mnSerde.micronaut.serde.jackson)
+
+    testImplementation('org.apache.groovy:groovy-json')
     testImplementation projects.micronautViewsSoy
     testImplementation projects.micronautViewsVelocity
+    testImplementation projects.micronautViewsHandlebars
+    testRuntimeOnly(mnLogging.logback.classic)
 }
 
 tasks.named('test') {
