+++ test-suite-groovy/src/test/groovy/io/micronaut/views/model/security/BooksController.groovy
@@ -1,0 +1,44 @@
+/*
+ * Copyright 2017-2019 original authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.micronaut.views.model.security;
+
+import io.micronaut.context.annotation.Requires;
+import io.micronaut.http.annotation.Controller;
+import io.micronaut.http.annotation.Get;
+import io.micronaut.security.annotation.Secured;
+import io.micronaut.security.rules.SecurityRule;
+import io.micronaut.views.View;
+
+import java.util.Arrays;
+import java.util.HashMap;
+import java.util.Map;
+
+@Requires(property = "spec.name", value = "SecurityViewModelProcessorSpec")
+//tag::class[]
+@Controller("/")
+public class BooksController {
+
+    @Secured(SecurityRule.IS_AUTHENTICATED)
+    @View("securitydecorator")
+    @Get
+    public Map<String, Object> index() {
+        Map<String, Object> model = new HashMap<>();
+        model.put("books", Arrays.asList("Developing Microservices"));
+        return model;
+    }
+}
+//end::class[]
