+++ test-suite-groovy/src/test/resources/views/fruits-processor.vm
@@ -1,0 +1,9 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <title>Fruits</title>
+</head>
+<body>
+<h1>config: $applicationName</h1>
+</body>
+</html>
