+++ test-suite-groovy/src/test/resources/views/securitydecorator.vm
@@ -1,0 +1,37 @@
+#*
+ * Copyright 2017-2019 original authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *#
+##tag::html[]
+<!DOCTYPE html>
+<html>
+<head>
+    <title>User Books</title>
+</head>
+<body>
+    #if( $security )
+      <h1>User: ${security['name']} email: ${security['attributes']['email']}</h1>
+    #end
+
+    #foreach($book in $books)
+      $book
+    #end
+
+    #if( $securitycustom )
+    <h1>Custom: ${securitycustom['name']}</span></h1>
+    #end
+
+</body>
+</html>
+##end::html[]
