+++ test-suite-groovy/src/test/resources/views/stark.hbs
@@ -1,0 +1,9 @@
+<!DOCTYPE html>
+<html>
+<head>
+<title>John Snow</title>
+</head>
+<body>
+<h1>John Snow</h1>
+</body>
+</html>
