+++ test-suite-groovy/src/test/resources/views/targaryen.vm
@@ -1,0 +1,9 @@
+<!DOCTYPE html>
+<html>
+<head>
+    <title>Aegon Targaryen</title>
+</head>
+<body>
+<h1>Aegon Targaryen</h1>
+</body>
+</html>
